Update libraries, separate sand_simulator module, little fixes to get zero clippy warnings
